Add Rubocop support
Rubocop checks for various kinds of problems, including
style issues.  We'll also include the performance checkers.

Railroader is a fork of Brakeman, and Brakeman didn't use Rubocop.
Thus, there are a huge number of issues normally reported by Rubocop
(this is a normal problem with "brown field" addition of a checker).
We will solve this by disabling most checks, and slowly enabling them
over time.

Disabling most checks still means that we will (from now on)
detect when we create an entirely new kind of issue.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>