Fixes for problems from updated dependencies
This fixes various problems caused by updates in the
components we depend on.

The ERB parser is now pickier, requiring a fix in
lib/railroader/parsers/template_parser.rb. That's fair,
what it was doing was wrong in the first place.  For more info, see:
https://redmine.ruby-lang.org/issues/15294

Unfortunately, this commit modifies gem_common.rb
so that sexp_processor must be < 4.13. That's undesirable.
I modified lib/ruby_parser/bm_sexp.rb to address issues
and allow the use of 4.13+, but it wasn't enough,
and it's more important that it *work*.

After this there is still one error on "bundle exec rake test":
CommandlineTests#test_exit_on_warn_default:
CLExit: CLExit
    test/tests/commandline.rb:16:in `quit'
    lib/railroader/commandline.rb:121:in `regular_report'
    lib/railroader/commandline.rb:142:in `run_report'
    lib/railroader/commandline.rb:35:in `run'
    lib/railroader/commandline.rb:20:in `start'
    test/tests/commandline.rb:52:in `cl_with_options'
    test/tests/commandline.rb:58:in `block in scan_app'

In the future we probably need to change the install approach.
There's no reason that Railroader's dependencies should have anything
to do with the code it analyzes.  Unfortunately "gem install"
does not honor Gemfile.lock, so we'll need to do things differently.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>